Everything as float seems to work

+++ flight_data_recorder/src/main.rs
@@ -204,28 +204,28 @@ define_sensors! {
     absolute_time <- "ABSOLUTE TIME" in Seconds as Integer;
 
     // CONTROL INFORMATION
-    yoke_x_position <- "YOKE X POSITION" in Position as Integer;
-    yoke_y_position <- "YOKE Y POSITION" in Position as Integer;
+    yoke_x_position <- "YOKE X POSITION" in Position as Float;
+    yoke_y_position <- "YOKE Y POSITION" in Position as Float;
 
-    yoke_x_position_ap <- "YOKE X POSITION WITH AP" in Position as Integer;
-    yoke_y_position_ap <- "YOKE Y POSITION WITH AP" in Position as Integer;
+    yoke_x_position_ap <- "YOKE X POSITION WITH AP" in Position as Float;
+    yoke_y_position_ap <- "YOKE Y POSITION WITH AP" in Position as Float;
 
-    yoke_x_indicator <- "YOKE X INDICATOR" in Position as Integer;
-    yoke_y_indicator <- "YOKE Y INDICATOR" in Position as Integer;
+    yoke_x_indicator <- "YOKE X INDICATOR" in Position as Float;
+    yoke_y_indicator <- "YOKE Y INDICATOR" in Position as Float;
 
-    rudder_position <- "RUDDER POSITION" in Position as Integer;
-    rudder_pedal_position <- "RUDDER PEDAL POSITION" in Position as Integer;
-    rudder_pedal_indicator <- "RUDDER PEDAL INDICATOR" in Position as Integer;
+    rudder_position <- "RUDDER POSITION" in Position as Float;
+    rudder_pedal_position <- "RUDDER PEDAL POSITION" in Position as Float;
+    rudder_pedal_indicator <- "RUDDER PEDAL INDICATOR" in Position as Float;
 
-    brake_right_position <- "BRAKE RIGHT POSITION" in Position as Integer;
-    brake_left_position <- "BRAKE LEFT POSITION" in Position as Integer;
+    brake_right_position <- "BRAKE RIGHT POSITION" in Position as Float;
+    brake_left_position <- "BRAKE LEFT POSITION" in Position as Float;
 
     throttle_level_position1 <- "GENERAL ENG THROTTLE LEVER POSITION:1" in Percent as Float;
     throttle_level_position2 <- "GENERAL ENG THROTTLE LEVER POSITION:2" in Percent as Float;
 
-    spoiler_handle_position <- "SPOILERS HANDLE POSITION" in Position as Integer;
-    spoiler_right_position <- "SPOILERS RIGHT POSITION" in Percent as Integer;
-    spoiler_left_position <- "SPOILERS LEFT POSITION" in Percent as Integer;
+    spoiler_handle_position <- "SPOILERS HANDLE POSITION" in Position as Float;
+    spoiler_right_position <- "SPOILERS RIGHT POSITION" in Percent as Float;
+    spoiler_left_position <- "SPOILERS LEFT POSITION" in Percent as Float;
 
     flaps_handle_index <- "FLAPS HANDLE INDEX" in Number as Float;
     flaps_num_handle_positions <- "FLAPS NUM HANDLE POSITIONS" in Number as Float;
